Split Bitcoin Core RPC client from fetcher
